fix(local): resolve Fn::Join Code.ImageUri for lambda.DockerImageCode.fromEcr (#286 Gap 2)
`lambda-resolver.ts:extractImageUri` only recognized flat strings and
`Fn::Sub`; `lambda.DockerImageCode.fromEcr(repo, { tagOrDigest })`
synthesizes the canonical CDK 2.x `Fn::Join` shape (verified via
`cdk synth` + `jq` against a real CDK 2.x app — same shape PR #280
handled for ECS `ContainerImage.fromEcrRepository`).

Pre-fix the resolver returned undefined for `Fn::Join` and fell through
to the ZIP-branch's misleading "no Runtime property" error.

Extract `tryResolveImageFnJoin` + supporting helpers
(`findEcrRepositoryRefInTree`, `resolveImageIntrinsic`,
`resolveImageIntrinsicAny`, `substituteImagePlaceholders`) and the
`FnJoinResolveOutcome` type from `ecs-task-resolver.ts` into the new
shared module `src/local/intrinsic-image.ts`. Both the ECS resolver
and the Lambda container resolver now import the same helper so a
future change to the canonical CDK shape gets fixed once. The
~400-line duplicate in `ecs-task-resolver.ts` is gone; the ECS module
re-exports `EcsImageResolutionContext` as a type alias for backward
compatibility with `src/cli/commands/local-run-task.ts`.

Resolution outcomes from `extractImageUri` for `Fn::Join` shapes:
- literal-only joins resolve cleanly (no state needed);
- same-stack ECR Repository refs return `needs-state` and surface a
  "deploy first / use fromImageAsset / pin a public image" hint;
- imported-repo shapes that reference `Ref: AWS::URLSuffix` without
  pseudo-parameter context surface a separate "pseudo parameter not
  resolved" hint (rather than the misleading ZIP-branch error);
- malformed shapes surface `unsupported-join` with a precise reason.

Tests: 14 new in `tests/unit/local/intrinsic-image.test.ts` covering
every outcome branch + Fn::Sub placeholder substitution; 4 new in
`tests/unit/local/lambda-resolver.test.ts` covering same-stack reject
/ malformed reject / imported-shape reject / literal-only resolve;
47 existing ECS resolver tests pass unchanged against the shared
helper. Full suite: 3004/3004.
